runtime/Dockerfile: build a lean runtime container
First steps towards building a lean runtime container. The starting
point is the bare-bones ubi-micro image.

Using a two-step build, ubi-minimal as a builder image, install the
known prerequisite system dependencies. So far these are universal
dependencies: libstdc++ and zlib, but we also need to tackle specific
module dependencies.

Much of this is very specific to the spring boot example.